requirements.txt
Flask
Flask-JWT-Extended
Flask-PyMongo
cryptography
pytest
Flask-Bcrypt
